+++ NEWS.md
@@ -1,6 +1,8 @@
 Version 0.5.1-dev
 2012-??-??
 
+ - Automatically localise external resources in CartoCSS (#4)
+
 Version 0.5.0
 2012-09-03